# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_handler.importCancelledResponse:
    properties:
      cancelled:
        type: boolean
    type: object
  internal_handler.importStartedResponse:
    properties:
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a modern RSS reader API.
  title: Gist API
  version: "1.0"
paths:
...
  /opml/import:
    delete:
      description: Cancel the current import task
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.importCancelledResponse'
      summary: Cancel Import
      tags:
      - opml
    post:
      consumes:
      - multipart/form-data
      - text/xml
      description: Start importing feeds and folders from an OPML file
      parameters:
      - description: OPML file to import
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.importStartedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Import OPML
      tags:
      - opml
...
swagger: "2.0"